U(TestService) them verify test

+++ src/main/java/com/englishweb/h2t_backside/service/test/impl/SubmitToeicServiceImpl.java
@@ -87,7 +87,7 @@ public class SubmitToeicServiceImpl extends BaseServiceImpl<SubmitToeicDTO, Subm
         return Double.valueOf(submits.stream()
                 .max(Comparator.comparing(SubmitToeic::getCreatedAt))
                 .map(SubmitToeic::getScore)
-                .orElse(0));
+                .orElse((int) 0));
     }
 
     @Override
